fix problem where a child BgThread of a BgThread can't get the RootIsolateToken.instance
in spawned Isolates, apparently the RootIsolateToken.instance value is null, so it has to be
passed along from parent to child to child to child, etc as a chain.

this does that by sticking a new static variable in BgThread that is set BEFORE Isolate.spawn()
so that a child thread will inherit a copy, that it can then use instead of RootIsolateToken.instance
for creating ITS children BgThreads.

it's a global value anyhow, so it doesn't really matter that its static.

This way of using CoW copy on write lets me avoid having to pass the token around between
thread object and application code, since the dispatcher is a static method and object lifespans
are pretty complicated already.

it works, too.  :-)